drbd: Fix a node node declaring a peer as Outdated after initial resyncs
This was interesting. The observation is that after the initial resync
one of the nodes wrongly sees one of the peers as 'Outdated'.

Now here is a walk through of the events:

This is a 3 node scenario. At the start, node n2 has the newest data.
Nodes n3 and n4 are connecting to n2 and to each other. From the perspective
of n3, we are connecting to the other nodes nearly concurrently:

[Thu Jul  9 10:48:14 2020] drbd res/0 drbd1000 n2: drbd_sync_handshake:
[Thu Jul  9 10:48:14 2020] drbd res/0 drbd1000 n2: self 0000000000000004:0000000000000000:A351D57F60A5F2D2:0000000000000000 bits:0 flags:24
[Thu Jul  9 10:48:14 2020] drbd res/0 drbd1000 n2: peer D20E1D7807C4711E:A351D57F60A5F2D2:0000000000000000:0000000000000000 bits:1536 flags:20
[Thu Jul  9 10:48:14 2020] drbd res/0 drbd1000 n2: uuid_compare()=target-set-bitmap by rule 20
[Thu Jul  9 10:48:14 2020] drbd res/0 drbd1000 n2: Setting and writing the whole bitmap, fresh node
[Thu Jul  9 10:48:14 2020] drbd res/0 drbd1000 n4: drbd_sync_handshake:
[Thu Jul  9 10:48:14 2020] drbd res/0 drbd1000 n4: self 0000000000000004:0000000000000000:A351D57F60A5F2D2:0000000000000000 bits:0 flags:24
[Thu Jul  9 10:48:14 2020] drbd res/0 drbd1000 n4: peer 0000000000000004:0000000000000000:0000000000000000:0000000000000000 bits:0 flags:24
[Thu Jul  9 10:48:14 2020] drbd res/0 drbd1000 n4: uuid_compare()=no-sync by rule 10
[Thu Jul  9 10:48:14 2020] drbd res/0 drbd1000 n4: bitmap content (13300 bits)

We begin a resync from n2. No resync with n4,
since both of us have the same current UUID.

Then an update comes from n4 where it announces that it is D_UP_TO_DATE now.
(Since it finished its resync from n2.)

[Thu Jul  9 10:48:14 2020] drbd res/0 drbd1000 n4: pdsk( DUnknown -> Inconsistent ) repl( Off -> Established )
[Thu Jul  9 10:48:14 2020] drbd res/0 drbd1000 n4: pdsk( Inconsistent -> UpToDate )

Then we finish our resync from n2:
[Thu Jul  9 10:48:16 2020] drbd res/0 drbd1000 n2: Resync done (total 1 sec; paused 0 sec; 53200 K/sec)
[Thu Jul  9 10:48:16 2020] drbd res/0 drbd1000 n2: updated UUIDs D20E1D7807C4711E:0000000000000000:A351D57F60A5F2D2:0000000000000000
[Thu Jul  9 10:48:16 2020] drbd res/0 drbd1000: disk( Inconsistent -> UpToDate )
[Thu Jul  9 10:48:16 2020] drbd res/0 drbd1000 n4: pdsk( UpToDate -> Outdated ) resync-susp( connection dependency -> no )
[Thu Jul  9 10:48:16 2020] drbd res/0 drbd1000 n2: repl( SyncTarget -> Established )

We set the peer n4 to Outdated because we think it has an older UUID.

This occurs because n4 finished the resync from n2 before
our disk state became visible:

[Thu Jul  9 10:48:14 2020] drbd res/0 drbd1000 n3: drbd_sync_handshake:
[Thu Jul  9 10:48:14 2020] drbd res/0 drbd1000 n3: self 0000000000000004:0000000000000000:A351D57F60A5F2D2:0000000000000000 bits:0 flags:4
[Thu Jul  9 10:48:14 2020] drbd res/0 drbd1000 n3: peer 0000000000000004:0000000000000000:0000000000000000:0000000000000000 bits:0 flags:24
[Thu Jul  9 10:48:14 2020] drbd res/0 drbd1000 n3: uuid_compare()=no-sync by rule 10
...
[Thu Jul  9 10:48:14 2020] drbd res/0 drbd1000 n2: Resync done (total 1 sec; paused 0 sec; 53200 K/sec)
[Thu Jul  9 10:48:14 2020] drbd res/0 drbd1000 n2: updated UUIDs D20E1D7807C4711E:0000000000000000:A351D57F60A5F2D2:0000000000000000
[Thu Jul  9 10:48:14 2020] drbd res/0 drbd1000: disk( Inconsistent -> UpToDate )
[Thu Jul  9 10:48:14 2020] drbd res/0 drbd1000 n3: resync-susp( connection dependency -> no )
[Thu Jul  9 10:48:14 2020] drbd res/0 drbd1000 n2: repl( SyncTarget -> Established )
[Thu Jul  9 10:48:14 2020] drbd res/0 drbd1000 n2: helper command: /sbin/drbdadm after-resync-target
[Thu Jul  9 10:48:14 2020] drbd res/0 drbd1000 n2: helper command: /sbin/drbdadm after-resync-target exit code 0
[Thu Jul  9 10:48:14 2020] drbd res n3: Preparing remote state change 2707118130
[Thu Jul  9 10:48:14 2020] drbd res n3: Committing remote state change 2707118130 (primary_nodes=1)
[Thu Jul  9 10:48:14 2020] drbd res n3: conn( Connecting -> Connected ) peer( Unknown -> Secondary )
[Thu Jul  9 10:48:14 2020] drbd res/0 drbd1000 n3: pdsk( DUnknown -> Inconsistent ) repl( Off -> Established )

Notice that the disk state of n3 becomes visible long after the resync finishes.

The code was:

		if (resync_finished && peer_disk_state[NEW] != D_UNKNOWN)
			drbd_send_uuids(peer_device, 0, 0);

The "peer_disk_state[NEW] != D_UNKNOWN" was here to avoid interference with
sending the initial state. Probably the change from April 2020 where the
initial state changes got merged into a single one, made this bug more likely
to trigger.

The problem is that when a resync finishes with one peer while we are
establishing a connection with another peer it never sends the updated UUID to
the newly joining peer.
"While we are establishing" is between sending my initial UUID and reaching
  repl_state > L_OFF (or peer_disk_state[NEW] != D_UNKNOWN)

The fix is to send a UUID packet again in the case where our current UUID
changed between sending the initial one and reaching repl_state established.