drbd: Fix race condition leading to inconsistent resync decisions
In disk_states_to_goodness() the local and remote disk states get compared.
The local disk state might change between sending it to the peer and arriving
in disk_states_to_goodness() function. If that happens the nodes will find
inconsistent resync decisions.

This is only relevant if the current UUIDs are equal and we look at the disk
states in addition. We got log-files that showed a 3 node setup, two nodes
starting a resync from a common source node concurrently. The issue described
above happened between the two resync target nodes.

Fix that by using the last disk state we sent to the peer instead of the
current local disk state.