drbd: Do not fail read requests from paused resync targets
instead send a cancel reply if we can not provide the data
in this moment.

It is very likely that these resync requests where already
in flight when the resync was paused. Failing the requests
means that the whole resync operation will leave the resync
target in inconsistent state.

Instead, cancel those requests, they will be re-issues by
the resync target after the pause.