ci: update virter and vmshed to v0.5.0
The main reason for the virter upgrade is to get a fix for a regression
that caused exit codes from docker provisioning steps to not be
propagated correctly.

Also, this gives us virter's new static DHCP mode feature, which the new
vmshed can use to improve the situation surrounding DHCP leases.