drbd: Fix sometimes Outdated disk state after a primary crash and rejoin
When a primary node crashes, one of the existing nodes gets promoted
to primary and then the former primary comes back (as secondary).

Now, if the newly promoted primary does not issue a single write request
it is on the same current UUID as all the other nodes.

  The UUID comparison code has the idea that after a crash of a primary
  node we need to do a resync (that is reasonable under the assumption
  that some IO was in flight before/at the time of the crash, and that
  there are set bits in the bitmap).

So, we get resyncs between the two surviving noded and the rejoining crashed
node. Initially, the crashed primary is considered to become resync source,
but later one the current primary is selected as sync-source. So the
rejoining node is sync-source for one peer and sync-target for its other
peer concurrently.

If the amount to resync is small (or even empty) all that happens so quickly
that the node at the end of the resync chain never pauses its resync. When that
happens its becomes Outdated at the end of its resync.

The fix: When all that happens and the sync-target node notices that it
becomes Outdated at the end of a resync, then it tries to get a resync
from the primary neighbor.
The challenge is that it and the primary are at the same current-UUID
(remember, there where no writes after the crash of the former primary
and the promotion of the current primary).
As solution, I decided to implement "inheriting" of the CRASHED_PRIMARY
falg. It is inherited from the sync-source with a resync that does
not end in an UpToDate disk.