drbd: improve workflow for adding new volumes to active resources
To initialize a new volume, we document and use
  new-current-uuid --clear-bitmap # to skip the initial sync
  new-current-uuid --force-resync # to force a resync

But some code paths check the current uuid for == UUID_JUST_CREATED,
which is currently not true for a resource in Primary role,
on a Primary it becomes (UUID_JUST_CREATED | UUID_PRIMARY).

This minimal fix, to *not* |= UUID_PRIMARY, even if we are primary,
if the current uuid == UUID_JUST_CREATED, allows "new-current-uuid"
to work on "just created" volumes added at runtime to an active resource.

HOW IT WORKS:

 - have a configured and healthy resource with one or more volumes
   connected to one or more peers
 - use it (have it Primary on one node)
 - add volume to resource config
 - create-md
 - adjust (on all! peers; creating a new volume on a connected resource
   will wait for that volume to appear on all connected peers as well)

That was already possible, but the volume will now be
"Inconsistent" on all peers (because it was just created).

With this patch, now you can choose to:

skip the initial sync
---------------------

Because you know the data is already identical everywhere, or because
you do not care now, and will immediately follow-up with a "mkfs" or
equivalent. On any node, do:

  new-current-uuid --clear-bitmap res/vnr

[Without this patch, that would fail on a Primary,
and appear to succeed on a Secondary, even though
it leaves the disk state on the Primary in "Inconsistent"]

force an initial resync
-----------------------

On the node that shall become SyncSource, do

  new-current-uuid --force-resync res/vnr

Yes, that may even be a node that is currently in Secondary role, and
initiate the initial resync from there to the other peers, all of which
will transition to SyncSource, including the node that is currently in
Primary role.

[Without this patch, that would simply fail]