drbd: Serialize connects and promotes properly
We see in logs that it can happen that two nodes that are establishing a
connection exchange their initial states and UUIDs, but before they can
change their connection states to connected, an other two-pc (usually
a promote) comes first.
I got logs where the promote was on one of the two nodes establishing
the connection, but it could happen that the promote is executed on
a third node.

After tinking about it for hours, and following the wrong ideas for
too long, I wonder why I did not came up with this elegant solution
a long time ago.

We have a two-phase-commit for establishing the connection. Simply
send the initial state after we did the prepare of the two-phase-commit.
The initiator of the two-phase-commit then receives the initial states
of all the others and starts the commit-phase after that.
the other node receives the initial states from its peer and after that
the commit packet. It does not need to wait for anything explicitly.

The two-phase-commit code ensures proper serialization of two-phase
commits, therefore no promote can now go in between as it happened
before.

Regarding cleanup_remote_state_change(). Starting with this change we have
two-phase-commits that start while the connection state is still
C_CONNECTING. We need to consider this in any_connection_up().

A possible improvement for the future: Generalize the idea that a
two-phase-commit can have code that is executed between prepare and
commit.