drbd: Reliably detect split-brain on both sides
So far two nodes simply sent the P_UUIDS110 packet towards each other. When one
detected that there is split-brain it immediatly disconnects. The other node
might notice the disconnect on the control socket first and misses to detect
the split-brain situation.

With the previous commit, we got proper two-phase-commits for establishing
the connection. Leveraging on that we allow both nodes to detect a split-brain
and do the disconnect when committing the new state on both sides.

A part of this change is when the meta-connection is shut down by the peer,
we always wait if the cstate goes below C_CONNECING (changed from C_CONNECTED)
soon.