ci: fix CI build for DRBD 10
DRBD 10 excludes the release part from the "REL_VERSION", which is the
version used in the name of the source tarball. Adjust the CI build to
match.

These changes have been made in such a way that they can be used largely
the same on the drbd-9.0 branch to reduce merging effort.