drbd: fail RW open early, if someone set us RO using blockdev --setro
In case someone wants to read-only share a DRBD to all nodes, we can
help them do that by allowing them to explicitly set us "read-only" in
the block device layer, and check that flag during open.

Todo: maybe catch that ioctl, and propagate it to all nodes.
Or add resource option "read-only", so we don't have to deal with
ambiguities during handshakes.