drbd: Fix rcu_read_lock() for online resizing
The callgraph is:

drbd_adm_resize()
  change_cluster_wide_device_size()
    complete_remote_state_change()
      begin_remote_state_change()
        rcu_read_unlock()

The purpose of complete_remote_state_change() is:
'Wait for other remote state changes to complete'

In case of contention, it assumes that it has to release the rcu_read_lock.
So, better hold the rcu_read_lock around calling complete_remote_state_change().