drbd: Fix unexpected NetworkFailure states
The following can happen when a 'drbdsetup disconnect' is followed by a
'drbdsetup connect' quickly.

In finish_state_chage() t_state of the receiver thread is set to EXITING
before the new state is assigned to the [NOW] members, i.e. before it
becomes visible (without locking). In addition to that, a signal is delivered
to the receiver thread.

When the receiver thread was trying to establish a connection it will
return from transport->ops->connect() in the conn_connect() function.
There it tests the state. When that is done without locking, it can
read the previous state there.

The effect of this bug looks like this in the logs:

1 res d2: conn( Connecting -> Disconnecting )
  res d2: Restarting sender thread
  res d2: Connection closed
  res d2: helper command: /sbin/drbdadm disconnected
  res d2: helper command: /sbin/drbdadm disconnected exit code 0
2 res d2: conn( Disconnecting -> StandAlone )
3 res d2: conn( StandAlone -> Unconnected )
4 res d2: Restarting receiver thread (from drbd_w_resync-b [52167])
  res d2: conn( Unconnected -> Connecting )
  res d2: Handshake to peer 1 successful: Agreed network protocol version 118
  res d2: Feature flags enabled on protocol level: 0xf TRIM THIN_RESYNC WRITE_SAME WRITE_ZEROES.
  res d2: Starting ack_recv thread (from drbd_r_resync-b [52188])
5 res d2: conn( Connecting -> NetworkFailure )

**1
The call to 'drbdsetup disconnect' triggers the state change, t_state is
set to EXITING.

**2
The receiver reads the old state (Connecting), calls conn_disconnet() and
but retries to connect.

**3
A call to 'drbdsetup connect' causes this state change. The state engine
calls from w_after_state_change() drbd_thread_start(receiver).

**4
The drbd_thread_start() functions finds the t_state still being EXITING
and sets it to RESTARTING. The "Restring ... from" message is unique to
this situation.

**5
It got connected and enters the drbdd() packet receive loop. Here it notices
that the thread state is not the expected RUNNING, but RESTARTING. Therfore
it exits from this loop. Early in conn_disconnect() it does a state transition
to C_NETWORK_FAILURE (that has only effect if the state is not yet below
C_CONNECTED).