drbd: Sanitize the semantics of INITIAL_STATE_SENT
The name suggests that the bit is set after the initial state was sent.
Due to historical reasons, there was some code that cleared that bit in
various situations. The one that cleared it when the connection state is
lower than C_CONNECTED was problematic since the recent change e595bfc6.

That would need to change based on the agreed protocol version.
But all that is too complicated and no longer used. Simply remove
all that, and clear it before establishing a connection.
`git grep clear_bit.INITIAL_STATE_SENT` has now a single hit, and
that this exactly how it should be.

The visible symptom of this wrong early clearing of the INITIAL_STATE_SENT
bit was that the `resync-after-primary-crash` terminated leaving one
node in SyncSource and the other node in Established connection states.

That happened because one side does not set initial_handshake to
true in drbd_uuid_compare(), and misses to realize that its
CRASHED_PRIMARY was cleared due to a resync with another node.