compat: do use ->queuedata in compat implementations
Since recent commit a43a7af7e9a2 ("drbd: stop using ->queuedata"), the drbd
device is no longer stored in the queue's queuedata field, but only in the
private_data field of the bio->bi_disk.

However, in some of our compat patches (namely blk_queue_merge_bvec and
blk_queue_plugged), we do not have access to the bio. Thus, when we use
this compat implementation, also store the drbd_device in the queuedata
field so we can use it there.