drbd: Fix two-primaries connect behaviour
With the recent commit e595bfc6 ('Serialize connects and promotes properly')
I broke the behavior that both nodes go to StandAllone when both sides
are Primary and dual-primary is not allowed by the configuration.

The idiomatic way with the two-phase-commit based state transition to
Connected, is to react to a P_TWOPC_NO with ceasing connection attempts and
going to StandAlone.

With aborting a two-phase-commit that intended to transition from Connecting
to Connected, we need to clear the INITIAL_STATE_* bits when closing the socket.
Otherwise a left over INITIAL_STATE_SENT causes __is_valid_soft_transition()
to return SS_IN_TRANSIENT_STATE for every call.