drbd: remove ->this_bdev
Same idea as upstream Linux kernel commit 155bd9d1abd6
("drbd: remove ->this_bdev").

The changes from drbd_get_capacity(device->this_bdev) to
get_capacity(device->vdisk) obviously work fine for us, but we have a
passage of code downstream which is not in the upstream version, and
that is supposed to flush the page caches *above* us when we become
secondary:

    fsync_bdev(device->this_bdev);

Instead of doing this directly, we need to quickly open the bdev to then
be able to flush it and close it again.