drbd: notify peer in bitmap exchange of new out-of-sync sectors
If we are SyncTarget while exchanging bitmaps with another peer, we need
to inform that peer of blocks that become out-of-sync. They may or may
not be in the exchanged bitmap, but there is no harm in being
conservative and always sending the out-of-sync packet here.

This fixes a problem that was observed where a later sync missed a
block that was synced while the bitmaps were being exchanged.