drbd: send sync target UUID rather than current data UUID
When we are currently sync target, we are receiving newer data. We want
to resolve other syncs on the basis of this target UUID so that they
occur in the correct direction once the ongoing sync in finished.

This is relevant in this case, for instance:
* Node A has the newest data, nodes B and C have the same (consistent)
  data, all disconnected.
* A and B connect and start sync A->B.
* B and C connect.

Prior to this change, B and C would enter a sync C->B due to the fact
that C had a "better" disk state Outdated than B which was Inconsistent.
Then the sync A->B would be paused, because C->B had a smaller
difference. Once the sync C->B was complete, B considered itself to be
UpToDate even though it actually had older data.

With this change, B and C enter a paused sync B->C instead. Once the
sync A->B is complete, the sync B->C runs.

When this "resolved UUID" is the same as our current UUID it means that
we have the same data as the SyncSource of this peer. The flag
UUID_FLAG_SYNC_TARGET indicates that the peer's UUID is a "resolved
UUID" so that we can detect this. In this situation, we must be careful
not to use this "resolved UUID" as a reason to skip a sync or clear
bitmaps.  Instead we become (Paused)SyncSource to this peer based on the
disk states.