drbd: fix slow sync when sync requests are answered quickly
This fixes a slow sync which occurred when the following conditions were
met:
* The peer responds faster to resync data requests than c_max_rate
* The resync request-response cycle for a small request occurs so
  quickly that c_max_rate prevents even a single block being requested
  on the next iteration.

In that situation the sync followed a distinctive pattern. The first
iteration requested a fairly arbitrary number of blocks based on the
logic for the start of the sync. On each further iteration, c_max_rate
caused fewer blocks to be requested than on the previous iteration.
After a couple of iterations, the request was very small. This iteration
was so fast, that no blocks were requested on the next one. The
next iteration was scheduled for 0.1s (RS_MAKE_REQS_INTV) later.  The
process restarted.

Fix this by rescheduling sooner when the resync controller determines
that no blocks should be requested.

This causes us to end up following a different pattern. An iteration
requests a small number of blocks. The response is fast, so the next
iteration requests no blocks but reschedules soon. The iteration after
requests a small number of blocks again.