drbd: Pause resync if a sync-source's disk becomes inconsistent
In a three (or more) node resource with two ongoing resyncs, the
downstream resync waits for the upstream resync to finish first.

Now if the upstream resync aborts due to connection loss the
downstream resync should be paused (potentially for a long time) if the
resync target has consistent (well, acutally Outdated) data.

With that DRBD will _not_ destroy the maybe last consistent copy of the
data.