drbd: allow L_PAUSED_SYNC_T from a D_INCONSISTENT peer
A peer that is Inconsistent due to a sync from some other node will
present the target UUID of that sync (the "resolved UUID") as its own
current UUID. If our data is older than that, we should become
PausedSyncT until the peer has received the newer data, at which point
we can sync from it. Allow this intermediate state.