drbd: Delay marking inconsistent upon becoming paused-resync-target
When a node with an UpToDate or Consistent disk state becomes
paused-resync-target, delay the downgrade of the disk state. In the moment
the resync makes progress for the first time (goes into resync-target
state) the disk becomes marked as Inconsistent.

That might help if the paused-resync is aborted, before it ever resyncs
a single block.

The explicit marking as inconsistent in drbd_start_resync() were not
necessary since the sanitize_state() function applies the implications
of the replication state to the disk states.

Within sanitize_state() the mentioned functionality needs to move past
the point where it finds out if a resync will start immediately or if
it enters its paused incarnation first.