drbd: Disable detection of finished resyncs while bitmap is in use
drbd_uuid_detect_finished_resyncs() should not operate on bitmaps
and bitmap-base UUIDs when an operation like resync or online
verify is ogoing.
(With the excection with the peer from which we got the uuid-set
 update we are just processing)

It will be triggered later on again when that bitmap using operation
terminates and exchanges UUIDs again.