drbd: set resync-source's current-uuid as exposed-uuid when becoming resync-target
In case resync-target had D_UP_TO_DATE data, it gives it up in the moment it
starts the resync. Application reads will be served with the data from the
sync-source starting in this moment. So, we should reflect this in the
exposed-uuid.

This changes the behaviour of a promotion event on a resync target.
Consider a 3 node system (A, B, and C) with resyncs from A to B and
A to C running.

A promote on B caused through drbd_uuid_new_current() a call to
a_lost_peer_is_on_same_cur_uuid() returned true. B generated a new
current-uuid. Which is not necessary nor desired.

The second part of this commit makes sure that no new current-uuid
is generated when this other node is resync-target itself.