drbd: Combine nested 2PC local connect results during connect
The commit e595bfc635
'drbd: Serialize connects and promotes properly'

failed to consider that behind the node that accepts a connection might
be futher nodes. In context of a two-phase commit they are called
"nested nodes".

The central parts of this change are in cluster_wide_reply_ready() and
get_cluster_wide_reply(). From now on it waits until the local result
and the nested result are ready and get combined.

The same logic determining when and how to reply to a two-phase commit
is used on the initiator to determine when the two-phase commit is
complete and what its outcome was.

A (nested) two-phase commit completes as soon as a "retry" reply is
received. Retries may be sent immediately when a request is received
which conflicts with an ongoing two-phase commit. Hence the peer may not
reply with its state in this situation. In contrast, "no" replies are
only ever sent after the state in such a connection two-phase commit.

Two-phase-commits can be queued when there are multiple that were not
initiated by us. So the flags storing the result of the state processing
need to be in the connection. Otherwise they might be applied to the
wrong two-phase commit.

This also reverts the behavior when we cannot connect for various
reasons back to how it was in 9.0.25. For instance, when we cannot
connect due to having multiple primaries. Similarly, it fixes the
decision of whether to disconnect or retry according to the rr_conflict
configuration. The state handshake outcome is communicated via the flags
CONN_HANDSHAKE_DISCONNECT and CONN_HANDSHAKE_RETRY.

Co-developed-by: Joel Colledge <joel.colledge@linbit.com>