drbd: Extend the meaning of rr-conflict
The connection setting "rr-conflict retry" is used to express that we
should retry connecting if, when connecting, we realize that the current
role is incompatible with the outcome of the negotiated resync
direction.

Extend that so that this setting also influences the behaviour if during
connect two primaries would connect.

The default is "rr-conflict disconnect". This change changes the default
behaviour. From now on it will cause both sides to go StandAlone.

In case the connect-two-phase-commit it retried due to this setting,
it is retried one per second. The low level TCP connections stay
connected.