containers: check for usermode_helper
This adds a check that can be enabled (e.g., by the k8s operator) to
enforce that in container setups the usermode_helper is disabled.
Otherwise it might happen that people load the module on the host and
have the usermode_helper on (i.e., the default). Then the host kernel
tries to call drbdadm, which might be installed, which then does not
find the resource and things start to break.