drbd: fix a source of data-curroption in consider_sending_peers_in_sync()
Consider 3 nodes, A - B - C (no connection between A and C) with a resync from
B to C and A is primary. Every the number of out-of-sync-bits in a resync-extents
(128MiB) reaches zero DRBD schedules a work item to propagate the news about the
partial resync success between B and C to A.

The function consider_sending_peers_in_sync() checks if that extent in question
still has no bit set and then sends the information to A.

What was wrong so far is that when a write request gets processed on node A
and B while the consider_sending_peers_in_sync() is scheduled on B. Also the
P_BARRIER and P_BARRIER_ACK happen in the meantime. Exactly then B executes
consider_sending_peers_in_sync() and sends to A that the whole RS extent is
now in sync between B and C with a P_PEERS_IN_SYNC packet.

In the meantime A applied to its out-of-sync-bitmap that the last write did
not make it to node C. Just a moment later it gets the P_PEERS_IN_SYNC packet
and clears that bit again.

With that, an out-of-sync-bit from A to C is lost! -- That might appear as
data-corruption at a later point in time when the primary role moves to node C.

Fix that by locking out mirrored writes from A on B while executing
consider_sending_peers_in_sync(). If such writes are in flight, simply give
clearing the resync-extent on A.

Reviewed-by: Philipp Reisner <philipp.reisner@linbit.com>
Reviewed-by: Lars Ellenberg <lars.ellenberg@linbit.com>
Signed-off-by: Philipp Reisner <philipp.reisner@linbit.com>
Signed-off-by: Lars Ellenberg <lars.ellenberg@linbit.com>
Signed-off-by: ZhangDuan <duan.zhang@easystack.cn>