drbd: forbid connecting two partitions with a primary in each
Only the initiator of the connect two-phase-commit can do that.
It needs to inspect primary_nodes node mask after the prepare phase.
This change adds that in drbd_state.c

(Regarding the change in connect_work())
In case a connect two-phase-commit gets rejected, treat SS_TWO_PRIMARIES
in the same way as SS_CW_FAILED_BY_PEER, since the former is just the
local case of the other one.

As only the initiator of the connect can find out if it should be
retried or connection retries stopped, remove attempts to guess that
from process_twopc().  Replace that with a P_DISCONNECT packet sent from
the only place where the necessary information is available.