drbd: Do not mark more and more nodes as MDF_NODE_EXISTS
Since 2f144892fa 'drbd: Allow resync of a re-created peer from day0 UUID'
DRBD includes a bitmap-UUID of one unallocated bitmap-slot in the UUIDs
packet.

Accidentialy this change introduced that with each such P_UUIDS110
packet it received one more node gets marked as MDF_NODE_EXISTS. That
in turn can trigger that the drbd_uuid_detect_finished_resyncs()
function clears bitmap UUIDs, and that can cause the generation of
a new current UUID when it is not expected.

Fix that by telling the peer which of the bitmaps is the one
that is for an unallocated bitmap-slot.