drbd: hide kABI breakage check behind a cocci patch
Newer kernel versions forbid including vermagic.h from modules. We use
it for our RHEL kABI breakage double check. Since this is not really
supposed to be in upstream code anyways, remove it from drbd_main.c and
patch it back in when needed.

The "when needed" is defined as "kernels that allow modules to include
vermagic.h", which covers all kernels that might be affected by the kABI
breakage (and some more, but that doesn't hurt).