drbd: Restore compatibility with drbd-9.0.25 and earlier
With commit 50836ed637
 'Combine nested 2PC local connect results during connect'
we broke the compatibility with earlier protocols.

Only starting with protocol 118 we should wait for the outcome
of the handshake before completing the two phase commit.

In protcols < 117 the conn_connect2() and receive_state() happen
after the two-phase-commit that establishes the connection.

In protcol == 117 the conn_connect2() and receive_state() happen
before the two-phase-commit that establishes the connection.

Starting with protocol >= 118 it conn_connect2() and receive_state()
are executed as part of the two-phase-commit between prepare and
commit phase.