drbd: do not send a barrier to peers on REQ_PREFLUSH
As explained in the comments in the code, this is not necessary. It
hurts performance by decreasing the I/O depth on the peers. Remove it.

Lars Ellenberg who introduced this, says that it was most likely
intended as a way of implementing the "wait for barrier" logic for
protocol C peers that is mentioned in the comment. However, this was
never finished, leaving an unnecessary barrier.