drbd: fix corruption of pending_master_completion list
This fixes corruption of the device->pending_master_completion[WRITE]
list, whose entries are req->req_pending_master_completion. The issue
caused the following warning in kernels with list debugging enabled:
[11696.882946] WARNING: CPU: 0 PID: 4613 at lib/list_debug.c:59 __list_del_entry+0xa1/0xd0
[11696.882947] list_del corruption. prev->next should be ffff9898feefec78, but was ffff98983ce0d358
...
[11696.883013] Call Trace:
[11696.883017]  [<ffffffffb3d7ff85>] dump_stack+0x19/0x1b
[11696.883020]  [<ffffffffb369bd18>] __warn+0xd8/0x100
[11696.883024]  [<ffffffffb369bd9f>] warn_slowpath_fmt+0x5f/0x80
[11696.883027]  [<ffffffffb39a4ca1>] __list_del_entry+0xa1/0xd0
[11696.883037]  [<ffffffffc0b2ada0>] drbd_req_put_completion_ref+0x220/0x3f0 [drbd]
[11696.883047]  [<ffffffffc0b2b3ff>] mod_rq_state+0x48f/0x830 [drbd]
[11696.883057]  [<ffffffffc0b2ba39>] __req_mod+0x159/0x9d0 [drbd]
[11696.883066]  [<ffffffffc0b17e1b>] validate_req_change_req_state+0xbb/0x130 [drbd]
[11696.883075]  [<ffffffffc0b181dd>] got_BlockAck+0xcd/0x200 [drbd]
[11696.883085]  [<ffffffffc0b291c9>] drbd_ack_receiver+0x239/0x600 [drbd]

The cause of the issue was that the request was queued on
device->submit.writes before being added to
device->pending_master_completion. Hence, with unfortunate timing, it
would be added to device->pending_master_completion in
drbd_send_and_submit and then again in drbd_queue_write. This corrupted
the list.

The problem was introduced in
cd4288e83160 drbd: another fine-grained lock pending_completion_lock
and related commits removing req_lock.

Fix it by correcting the order we add the request to each of the lists.