drbd: let a second diskful node become primary with allow-two-primaries
The commit
187bddac03f7 drbd: forbid connecting two partitions with a primary in each
removed a condition allowing primaries to connect as long as they are
strongly connected. This prevented diskful primaries from connecting
when even with allow-two-primaries enabled.

In particular, this broke live migration between diskful nodes.

Restore that condition to fix it. Place it after the new
multiple_primaries_allowed check to ensure that we do not accidentally
allow multiple primaries when allow-two-primaries is disabled.