drbd: fix missing decrement of unacked counter on error
When clearing up on errors at the end of receive_Data, we need to
decrement the unacked count if it has been incremented. Otherwise the
corresponding peer request is destroyed and the counter is left one too
high.

This can occur, for instance, in the following situation:
* Protocol C is configured
* We are Secondary
* A peer is Primary and performing I/O
* We disconnect

If a write is waiting for an activity log extent, then this wait will be
canceled due to the disconnect, and this issue will occur.

Add in the missing decrement of the unacked counter.