drbd: Fix concurrent attach
When attaching a disk on a peer, the relevant event is when the peer's disk
state enters D_NEGOTIATING state.  It was working for attach operations on a
single peer, but failed with multiple attach operations starting
concurrently.  Entering D_NEGOTIATING disk state is coordinated by a
two-phase-commit, so only one node can do that at a time.

The other chunk just makes sure that in a concurrent attach never
a peer-disk state of D_CONSISTENT becomes visible (before the negotiation
phase finished properly)