drbd: Fix end-of-resync for diskless peers of a resync-target
When we finish a resnc as resync-target, we need to rotate the
previous UUID into the bitmap-base slots of our peers that need it.
So far the code missed that a peer that is diskless now might
had the same current_uuid as we are synced-up to with now.
Also look at the peer_disk_state to avoid this mistake.