drbd: push old current UUID into history
This fixes an issue which occasionally occurred since
82e799216bac drbd: avoid pushing peer bitmap for us into our history

In the specific case that was observed, the issue was as follows:
* Nodes A, B and C are Connected and UpToDate with UUID X; B is Primary
* The connection C -> B is lost
* B fences C, fence-peer helper returns 3 (peer is inconsistent or
  worse)
* B creates a new UUID Y and sends it to A, with no nodes marked as weak
* A updates its current UUID and its bitmap UUID for C to Y, losing all
  record of UUID X
* A and C perform a sync handshake
* C reports "unrelated-data" since it has current UUID X which A does
  not know; in the specific case observed, A correctly decided that it
  should be sync source, but that was only due to interactions with
  another node which affected its history

In general we cannot assume that we can overwrite our current UUID even
when that UUID is not any of our bitmap UUIDs. It might be that the new
UUID does not arrive at one of the other nodes.

Fix the issue by pushing our old current UUID into the history when we
receive a new one.