drbd: Fix hitting BUG() via double add_timer() under special timing
This happens if in between the call to drbd_queue_work() and the execution
of the work itme (connect_work()) the receiver calls del_connect_timer().
Then del_connect_timer() does not find the timer pending but a little
time later, both the receiver threadn and the work item call add_timer.

An example of a timing leading to the BUG()

1 The receiver thread calls conn_connect() and adds the timer in line 996:
  add_timer(&connection->connect_timer). Then drbd_receiver() calls drbdd()
2 Lets say, now the timer expires and queues connect_timer_work in
  connect_timer_fn()
3 The receiver thread returns from drbdd() e.g. connection goes down or
  drbdsetup disconnect & drbdsetup connect comes in. It calls
  conn_disconnect()-> del_connect_timer() it does not delete the timer since
  it is currently not active.
4 Now the work gets scheduled and executes connect_work() in that particular
  run the change_cstate() probably failed with SS_CONCURRENT_ST_CHG (or maybe
  with SS_HANDSHAKE_RETRY) so it arms the timer with add_timer().
5 Then the receiver thread calls conn_connect() again, and arms the timer
  again -> BUG()

Fix this by: using mod_timer() instead of add_timer()